backend/requirements.txt
fastapi==0.119.0
openai==2.4.0
Pillow==12.0.0
protobuf==6.33.0
pytesseract==0.3.13
python-dotenv==1.1.1
Requests==2.32.5
uvicorn==0.37.0
